refactor(website): update hasScrolled body class implementation; move to app level, tidy up
